move scheduler to an internal/scheduler package.
The otelcol component scheduler code will need to be exposed to multiple
packages, but it doesn't make sense to make it part of the public API.
